Moving files into types dir Co-authored-by: Alex Sanders <alex@sndrs.dev>
