+++ libs/@guardian/source-react-components/src/button/@types/SharedButtonProps.ts
@@ -1,6 +1,6 @@
 import type { ReactElement } from 'react';
-import type { Props } from '../@types/Props';
-import type { ThemeButton } from './theme';
+import type { Props } from '../../@types/Props';
+import type { ThemeButton } from '../theme';
 
 export type ButtonPriority = 'primary' | 'secondary' | 'tertiary' | 'subdued';
 export type IconSide = 'left' | 'right';
