+++ libs/@guardian/source-react-components/src/link/Link.tsx
@@ -1,8 +1,8 @@
 import type { EmotionJSX } from '@emotion/react/types/jsx-namespace';
 import type { AnchorHTMLAttributes } from 'react';
+import type { SharedLinkProps } from './@types/SharedLinkProps';
 import { linkContents } from './shared';
 import { linkStyles } from './styles';
-import type { SharedLinkProps } from './types';
 
 export interface LinkProps
 	extends AnchorHTMLAttributes<HTMLAnchorElement>,
