+++ libs/@guardian/source-react-components/src/tiles/styles.ts
@@ -1,7 +1,7 @@
 import type { SerializedStyles } from '@emotion/react';
 import { css } from '@emotion/react';
 import { from, space, until } from '@guardian/source-foundations';
-import type { Columns } from './types';
+import type { Columns } from './@types/Columns';
 
 const collapseUntilTilesSpacing = css`
 	display: block;
